零时提交

+++ parasol-tags/parasol-tags-provider/src/main/java/com/ginkgocap/parasol/tags/service/impl/TagSourcesServiceImpl.java
@@ -1,15 +1,25 @@
 package com.ginkgocap.parasol.tags.service.impl;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang3.ObjectUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ginkgocap.parasol.common.service.exception.BaseServiceException;
 import com.ginkgocap.parasol.common.service.impl.BaseService;
+import com.ginkgocap.parasol.tags.exception.TagServiceException;
 import com.ginkgocap.parasol.tags.exception.TagSourceServiceException;
+import com.ginkgocap.parasol.tags.model.Tag;
 import com.ginkgocap.parasol.tags.model.TagSource;
+import com.ginkgocap.parasol.tags.service.TagService;
 import com.ginkgocap.parasol.tags.service.TagSourceService;
 
 /**
@@ -23,11 +33,15 @@ import com.ginkgocap.parasol.tags.service.TagSourceService;
 @Service("tagSourceService")
 public class TagSourcesServiceImpl extends BaseService<TagSource> implements TagSourceService {
 
+	private static Logger logger = Logger.getLogger(TagSourcesServiceImpl.class);
 	// 一个资源下有那些TagSource
 	private static final String LIST_ID_APPID_SOURCEID_SOURCETYPE = "List_Id_AppId_SourceId_SourceType";
 	private static final String LIST_ID_APPID_TAGID = "List_Id_AppId_TagId";
 
 	private static int MAX_TAG = 20; // 一个资源下最多创建的标签数
+
+	@Autowired
+	private TagService tagService;
 
 	@Override
 	public Long createTagSource(TagSource source) throws TagSourceServiceException {
@@ -36,12 +50,36 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 		ServiceError.assertPopertiesIsNullForTagSource("userId");
 		ServiceError.assertPopertiesIsNullForTagSource("sourceId");
 		ServiceError.assertPopertiesIsNullForTagSource("sourceType");
+
+		// 检查Tag是否存在,系统标签也可以处理
+		try {
+			Tag tag = tagService.getTag(source.getUserId(), source.getTagId());
+			// TODO : 添加系统标签的处理
+			if (tag == null) {
+				throw new TagSourceServiceException(ServiceError.ERROR_NOT_FOUND, "The label [" + source.getTagId() + "] you want to add does not exist"); // 要添加的标签不存在！
+			}
+		} catch (TagServiceException tagException) {
+			tagException.printStackTrace(System.err);
+			throw new TagSourceServiceException(tagException);
+		}
+
 		int count = countTagSourcesByAppIdSourceIdSourceType(source.getAppId(), source.getSourceId(), source.getSourceType());
 		if (count >= MAX_TAG) {
 			throw new TagSourceServiceException(ServiceError.ERROR_TOO_MANY, "Can't create too TagSource tag， Max is " + MAX_TAG);
 		}
 
 		try {
+			if (count > 0) {
+				List<TagSource> tagSources = getTagSourcesByAppIdSourceIdSourceType(source.getAppId(), source.getSourceId(), source.getSourceType());
+				for (TagSource existTagSource : tagSources) {
+					if (existTagSource != null && ObjectUtils.equals(existTagSource.getTagId(), source.getTagId())) {
+						TagSourceServiceException te = new TagSourceServiceException(ServiceError.ERROR_OBJECT_EXIST, "The  " + source.getTagId()
+								+ " tag that you want to create already exists");
+						logger.error(te.getErrorCode() + " " + te.getMessage());
+						throw te;
+					}
+				}
+			}
 			return (Long) saveEntity(source);
 		} catch (BaseServiceException e) {
 			e.printStackTrace(System.err);
@@ -51,20 +89,20 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 	}
 
 	@Override
-	public boolean removeTagSource(Long appId, Long userId , Long tagSourceId) throws TagSourceServiceException {
+	public boolean removeTagSource(Long appId, Long userId, Long tagSourceId) throws TagSourceServiceException {
 		ServiceError.assertAppidIsNullForTagSource(appId);
 		ServiceError.assertTagSourceIdIsNullForTagSource(tagSourceId);
 		ServiceError.assertUserIdIsNullForTagSource(userId);
 		try {
-			TagSource tagSource= this.getTagSource(appId, tagSourceId);
+			TagSource tagSource = this.getTagSource(appId, tagSourceId);
 			if (tagSource != null) {
-				if (ObjectUtils.equals(tagSource.getUserId(), userId)) {
+				if (ObjectUtils.equals(tagSource.getUserId(), userId) && ObjectUtils.equals(tagSource.getAppId(), appId)) {
 					return this.deleteEntity(tagSourceId);
 				} else {
 					throw new TagSourceServiceException(ServiceError.ERROR_NOT_MYSELF, "Operation of the non own TagSource");
 				}
 			}
-			
+
 			return this.deleteEntity(tagSourceId);
 		} catch (BaseServiceException e) {
 			e.printStackTrace(System.err);
@@ -72,18 +110,31 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 		}
 	}
 
-	
 	@Override
 	public TagSource getTagSource(Long appId, Long tagSourceId) throws TagSourceServiceException {
 		ServiceError.assertAppidIsNullForTagSource(appId);
 		ServiceError.assertTagSourceIdIsNullForTagSource(tagSourceId);
 		try {
-			return this.getEntity(tagSourceId);
+			TagSource ts = this.getEntity(tagSourceId);
+			if (ts != null) {
+				Tag tag = tagService.getTag(ts.getUserId(), ts.getTagId());
+				if (tag != null) {
+					ts.setTagName(tag.getTagName());
+				} else {
+					logger.warn("don't find the tag by id : " + ts.getTagId());
+					return null; // Tag不存在的时候，直接返回标示TagSource也不存在了
+				}
+			}
+			return ts;
 		} catch (BaseServiceException e) {
+			e.printStackTrace(System.err);
+			throw new TagSourceServiceException(e);
+		} catch (TagServiceException e) {
 			e.printStackTrace(System.err);
 			throw new TagSourceServiceException(e);
 		}
 	}
+
 	@Override
 	public List<TagSource> getTagSourcesByAppIdSourceIdSourceType(Long appId, Long sourceId, Integer sourceType) throws TagSourceServiceException {
 		ServiceError.assertAppidIsNullForTagSource(appId);
@@ -91,8 +142,59 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 		ServiceError.assertTagSourceTypeIsNullForTagSource(sourceType);
 		try {
 			List<TagSource> tagSources = this.getEntitys(LIST_ID_APPID_SOURCEID_SOURCETYPE, appId, sourceId, sourceType);
-			return tagSources;
+			List<TagSource> result = new ArrayList<TagSource>();
+
+			if (CollectionUtils.isNotEmpty(tagSources)) {
+				// 按用户分组
+				Map<Long, List<Long>> userTagIdsMap = new HashMap<Long, List<Long>>();
+				for (TagSource tagSource : tagSources) {
+					if (tagSource != null) {
+						Long userId = tagSource.getUserId();
+						List<Long> tagIds = userTagIdsMap.get(userId);
+						if (tagIds == null) {
+							tagIds = new ArrayList<Long>();
+							userTagIdsMap.put(userId, tagIds);
+						}
+						tagIds.add(tagSource.getTagId());
+					}
+				}
+
+				Map<Long, String> tagMap = new HashMap<Long, String>();
+				// 开始查询每个用户的标签
+				if (MapUtils.isNotEmpty(userTagIdsMap)) {
+					for (Long userId : userTagIdsMap.keySet()) {
+						List<Long> tagIds = userTagIdsMap.get(userId);
+						if (CollectionUtils.isNotEmpty(tagIds)) {
+							List<Tag> tags = tagService.getTags(userId, tagIds);
+							for (Tag tag : tags) {
+								if (tag != null) {
+									tagMap.put(tag.getId(), tag.getTagName());
+								}
+							}
+						}
+					}
+				}
+
+				// 开始组装
+				if (MapUtils.isNotEmpty(tagMap)) {
+					for (TagSource tagSource : tagSources) {
+						if (tagSource != null) {
+							String tagName = tagMap.get(tagSource.getTagId());
+							if (StringUtils.isNotBlank(tagName)) {
+								tagSource.setTagName(tagName);
+								result.add(tagSource);
+							}
+						}
+					}
+				}
+
+			}
+
+			return result;
 		} catch (BaseServiceException e) {
+			e.printStackTrace(System.err);
+			throw new TagSourceServiceException(e);
+		} catch (TagServiceException e) {
 			e.printStackTrace(System.err);
 			throw new TagSourceServiceException(e);
 		}
@@ -144,7 +246,7 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 			if (count != null && count > 0) {
 				int eachNumber = 20;
 				int cycles = count / eachNumber; // 每次删除20个
-				int rem = count % eachNumber; 	 // 余数
+				int rem = count % eachNumber; // 余数
 				cycles = rem > 0 ? cycles + 1 : cycles;
 				for (int i = 0; i < cycles; i++) {
 					int start = i * eachNumber;
@@ -160,7 +262,5 @@ public class TagSourcesServiceImpl extends BaseService<TagSource> implements Tag
 
 		return true;
 	}
-
-
 
 }
